requirements.txt
asgiref==3.2.7
Django==3.0.6
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
psycopg2==2.8.5
Pillow==7.1.2
numpy==1.18.5
opencv-python==4.2.0.34